+++ RLSimion-Lib/DQN.h
@@ -27,11 +27,11 @@ protected:
 
 	size_t m_numberOfActions;
 	size_t m_numberOfStateFeatures;
-	std::vector<float> m_stateVector;
-	std::vector<float> m_actionValuePredictionVector;
+	std::vector<double> m_stateVector;
+	std::vector<double> m_actionValuePredictionVector;
 
-	std::vector<float> m_minibatchStateVector;
-	std::vector<float> m_minibatchActionValuePredictionVector;
+	std::vector<double> m_minibatchStateVector;
+	std::vector<double> m_minibatchActionValuePredictionVector;
 
 	int* m_pMinibatchChosenActionIndex;
