+++ tools/NNCreatorLib/Network.cpp
@@ -41,7 +41,7 @@ CNetwork CNetwork::load(string fileName, CNTK::DeviceDescriptor &device)
 	return result;
 }
 
-void CNetwork::train(std::unordered_map<std::string, std::vector<float>&>& inputDataMap, std::vector<float>& targetOutputData)
+void CNetwork::train(std::unordered_map<std::string, std::vector<double>&>& inputDataMap, std::vector<double>& targetOutputData)
 {
 	NDShape outputShape = getOutputsFunctionPtr().at(0)->Output().Shape();
 	ValuePtr outputSequence = CNTK::Value::CreateBatch(outputShape, targetOutputData, CNTK::DeviceDescriptor::CPUDevice());
@@ -59,7 +59,7 @@ void CNetwork::train(std::unordered_map<std::string, std::vector<float>&>& input
 	m_trainer->TrainMinibatch(arguments, DeviceDescriptor::CPUDevice());
 }
 
-void CNetwork::predict(std::unordered_map<std::string, std::vector<float>&>& inputDataMap, std::vector<float>& predictionData)
+void CNetwork::predict(std::unordered_map<std::string, std::vector<double>&>& inputDataMap, std::vector<double>& predictionData)
 {
 	FunctionPtr outputPtr = getOutputsFunctionPtr().at(0);
 	ValuePtr outputValue;
